requirements.txt
allure-pytest==2.8.5
allure-python-commons==2.8.5
Appium-Python-Client==0.47
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
idna==2.8
importlib-metadata==0.20
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.2
pytest==5.1.2
pywin32==225
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
xlwt==1.3.0
zipp==0.6.0
